hm9/requirements.txt
protobuf==3.6.1
python-memcached==1.59
six==1.15.0